+++ src/app/(root)/(routes)/items/components/ItemList/index.tsx
@@ -1,0 +1,3 @@
+import ItemList from './ItemList'
+
+export default ItemList
